+++ database/migrations/2017_10_17_215430_create_categoria_cuota_socials_table.php
@@ -19,7 +19,7 @@ class CreateCategoriaCuotaSocialsTable extends Migration
             $table->integer('id_organismo')->unsigned();
             $table->foreign('id_organismo')->references('id')->on('organismos');
             $table->integer('categoria');
-            $table->integer('valor');
+            $table->double('valor');
             $table->timestamps();
         });
     }
